add2

+++ caracteres.c
@@ -4,7 +4,7 @@
 
 enum caracter {a,b,c,d,e,f,g,h,i,j,k,l,m,n,o,p,q,r,s,t,u,v,w,x,y,z,CERO,UNO,DOS,TRES,CUATRO,CINCO,SEIS,SIETE,OCHO,NUEVE,ARR,ABJ,DER,IZQ,ESP};
 
-const int caracter_a[7][2] = {
+const float caracter_a[7][2] = {
 	{0, 0},
 	{0, 4},
 	{2, 6},
@@ -14,7 +14,7 @@ const int caracter_a[7][2] = {
 	{0, 2}
 };
 
-const int caracter_b[12][2] = {
+const float caracter_b[12][2] = {
 	{0, 3},
 	{0, 6},
 	{3, 6},
@@ -29,14 +29,14 @@ const int caracter_b[12][2] = {
 	{3, 3},
 };
 
-const int caracter_c[4][2] = {
+const float caracter_c[4][2] = {
 	{4, 0},
 	{0, 0},
 	{0, 6},
 	{4, 6},
 };
 
-const int caracter_d[7][2] = {
+const float caracter_d[7][2] = {
 	{0, 0},
 	{0, 6},
 	{2, 6},
@@ -46,7 +46,7 @@ const int caracter_d[7][2] = {
 	{0, 0}
 };
 
-const int caracter_e[7][2] = {
+const float caracter_e[7][2] = {
 	{4, 0},
 	{0, 0},
 	{0, 3},
@@ -56,7 +56,7 @@ const int caracter_e[7][2] = {
 	{4, 6},
 };
 
-const int caracter_f[6][2] = {
+const float caracter_f[6][2] = {
 	{0, 0},
 	{0, 3},
 	{3, 3},
@@ -65,7 +65,7 @@ const int caracter_f[6][2] = {
 	{4, 6},
 };
 
-const int caracter_g[7][2] = {
+const float caracter_g[7][2] = {
 	{2, 2},
 	{4, 2},
 	{4, 0},
@@ -75,7 +75,7 @@ const int caracter_g[7][2] = {
 	{4, 4},
 };
 
-const int caracter_h[6][2] = {
+const float caracter_h[6][2] = {
 	{0, 0},
 	{0, 6},
 	{0, 3},
@@ -84,7 +84,7 @@ const int caracter_h[6][2] = {
 	{4, 0},
 };
 
-const int caracter_i[6][2] = {
+const float caracter_i[6][2] = {
 	{0, 0},
 	{4, 0},
 	{2, 0},
@@ -93,14 +93,14 @@ const int caracter_i[6][2] = {
 	{0, 6},
 };
 
-const int caracter_j[4][2] = {
+const float caracter_j[4][2] = {
 	{0, 2},
 	{2, 0},
 	{4, 0},
 	{4, 6},
 };
 
-const int caracter_k[6][2] = {
+const float caracter_k[6][2] = {
 	{0, 0},
 	{0, 6},
 	{0, 3},
@@ -109,13 +109,13 @@ const int caracter_k[6][2] = {
 	{3, 0},
 };
 
-const int caracter_l[3][2] = {
+const float caracter_l[3][2] = {
 	{4, 0},
 	{0, 0},
 	{0, 6},
 };
 
-const int caracter_m[5][2] = {
+const float caracter_m[5][2] = {
 	{0, 0},
 	{0, 6},
 	{2, 4},
@@ -123,14 +123,14 @@ const int caracter_m[5][2] = {
 	{4, 0},
 };
 
-const int caracter_n[4][2] = {
+const float caracter_n[4][2] = {
 	{0, 0},
 	{0, 6},
 	{4, 0},
 	{4, 6},
 };
 
-const int caracter_o[5][2] = {
+const float caracter_o[5][2] = {
 	{4, 0},
 	{0, 0},
 	{0, 6},
@@ -138,7 +138,7 @@ const int caracter_o[5][2] = {
 	{4, 0},
 };
 
-const int caracter_p[5][2] = {
+const float caracter_p[5][2] = {
 	{0, 0},
 	{0, 6},
 	{4, 6},
@@ -146,7 +146,7 @@ const int caracter_p[5][2] = {
 	{0, 3},
 };
 
-const int caracter_q[9][2] = {
+const float caracter_q[9][2] = {
 	{2, 2},
 	{3, 1},
 	{2, 0},
@@ -158,7 +158,7 @@ const int caracter_q[9][2] = {
 	{4, 0},
 };
 
-const int caracter_r[7][2] = {
+const float caracter_r[7][2] = {
 	{0, 0},
 	{0, 6},
 	{4, 6},
@@ -168,7 +168,7 @@ const int caracter_r[7][2] = {
 	{4, 0},
 };
 
-const int caracter_s[6][2] = {
+const float caracter_s[6][2] = {
 	{0, 0},
 	{4, 0},
 	{4, 3},
@@ -177,27 +177,27 @@ const int caracter_s[6][2] = {
 	{4, 6},
 };
 
-const int caracter_t[4][2] = {
+const float caracter_t[4][2] = {
 	{2, 0},
 	{2, 6},
 	{4, 6},
 	{0, 6},
 };
 
-const int caracter_u[4][2] = {
+const float caracter_u[4][2] = {
 	{0, 6},
 	{0, 0},
 	{4, 0},
 	{4, 6},
 };
 
-const int caracter_v[3][2] = {
+const float caracter_v[3][2] = {
 	{0, 6},
 	{2, 0},
 	{4, 6},
 };
 
-const int caracter_w[5][2] = {
+const float caracter_w[5][2] = {
 	{0, 6},
 	{0, 0},
 	{2, 2},
@@ -205,7 +205,7 @@ const int caracter_w[5][2] = {
 	{4, 6},
 };
 
-const int caracter_x[5][2] = {
+const float caracter_x[5][2] = {
 	{0, 0},
 	{4, 6},
 	{2, 3},
@@ -213,7 +213,7 @@ const int caracter_x[5][2] = {
 	{4, 0},
 };
 
-const int caracter_y[5][2] = {
+const float caracter_y[5][2] = {
 	{0, 6},
 	{2, 4},
 	{4, 6},
@@ -221,14 +221,14 @@ const int caracter_y[5][2] = {
 	{2, 0},
 };
 
-const int caracter_z[4][2] = {
+const float caracter_z[4][2] = {
 	{0, 6},
 	{4, 6},
 	{0, 0},
 	{4, 0},
 };
 
-const int caracter_0[5][2] = {
+const float caracter_0[5][2] = {
 	{4, 0},
 	{0, 0},
 	{0, 6},
@@ -236,12 +236,12 @@ const int caracter_0[5][2] = {
 	{4, 0},
 };
 
-const int caracter_1[2][2] = {
+const float caracter_1[2][2] = {
 	{2, 0},
 	{2, 6},
 };
 
-const int caracter_2[6][2] = {
+const float caracter_2[6][2] = {
 	{0, 6},
 	{4, 6},
 	{4, 3},
@@ -250,7 +250,7 @@ const int caracter_2[6][2] = {
 	{4, 0},
 };
 
-const int caracter_3[7][2] = {
+const float caracter_3[7][2] = {
 	{0, 0},
 	{4, 0},
 	{4, 3},
@@ -260,7 +260,7 @@ const int caracter_3[7][2] = {
 	{0, 6},
 };
 
-const int caracter_4[5][2] = {
+const float caracter_4[5][2] = {
 	{0, 6},
 	{0, 3},
 	{4, 3},
@@ -268,7 +268,7 @@ const int caracter_4[5][2] = {
 	{4, 0},
 };
 
-const int caracter_5[6][2] = {
+const float caracter_5[6][2] = {
 	{0, 0},
 	{4, 0},
 	{4, 3},
@@ -277,7 +277,7 @@ const int caracter_5[6][2] = {
 	{4, 6},
 };
 
-const int caracter_6[5][2] = {
+const float caracter_6[5][2] = {
 	{0, 6},
 	{0, 0},
 	{4, 0},
@@ -285,13 +285,13 @@ const int caracter_6[5][2] = {
 	{0, 3},
 };
 
-const int caracter_7[3][2] = {
+const float caracter_7[3][2] = {
 	{0, 6},
 	{4, 6},
 	{4, 0},
 };
 
-const int caracter_8[7][2] = {
+const float caracter_8[7][2] = {
 	{0, 3},
 	{0, 6},
 	{4, 6},
@@ -301,7 +301,7 @@ const int caracter_8[7][2] = {
 	{4, 3},
 };
 
-const int caracter_9[5][2] = {
+const float caracter_9[5][2] = {
 	{4, 0},
 	{4, 6},
 	{0, 6},
@@ -309,7 +309,7 @@ const int caracter_9[5][2] = {
 	{4, 3},
 };
 
-const int caracter_derecha[5][2] = {
+const float caracter_derecha[5][2] = {
 	{0, 3},
 	{12, 3},
 	{9, 5},
@@ -317,7 +317,7 @@ const int caracter_derecha[5][2] = {
 	{12, 3},
 };
 
-const int caracter_izquierda[5][2] = {
+const float caracter_izquierda[5][2] = {
 	{12, 3},
 	{0, 3},
 	{3, 5},
@@ -325,7 +325,7 @@ const int caracter_izquierda[5][2] = {
 	{0, 3},
 };
 
-const int caracter_arriba[5][2] = {
+const float caracter_arriba[5][2] = {
 	{5, -2},
 	{5, 10},
 	{3, 7},
@@ -333,7 +333,7 @@ const int caracter_arriba[5][2] = {
 	{5, 10},
 };
 
-const int caracter_abajo[5][2] = {
+const float caracter_abajo[5][2] = {
 	{5, 10},
 	{5, -2},
 	{3, 1},
@@ -341,7 +341,7 @@ const int caracter_abajo[5][2] = {
 	{5, -2},
 };
 
-const int caracter_espacio[1][2] = {
+const float caracter_espacio[1][2] = {
 	{0, 0},
 };
 
@@ -392,7 +392,7 @@ const char letras[]= {
 	[ESP]=' ',
 };
 
-const int (*caracteres[])[2]= {
+const float (*caracteres[])[2]= {
 	[a]=caracter_a,
 	[b]=caracter_b,
 	[c]=caracter_c,
